fix(session): ensure new chatbot page uses server session
